+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.1] - 2019-02-06
 ### Changed
 - Handle brace blocks on commands properly.
 - Break parent and return `do` blocks when called from a `command` node.
@@ -42,7 +44,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - Initial release 🎉
 
-[Unreleased]: https://github.com/CultureHQ/add-to-calendar/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/CultureHQ/add-to-calendar/compare/0.2.1...HEAD
+[0.2.1]: https://github.com/CultureHQ/add-to-calendar/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/CultureHQ/add-to-calendar/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/CultureHQ/add-to-calendar/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/CultureHQ/add-to-calendar/compare/v0.1.0...v0.1.1